调整BuildMethod和NotGrpcMethodAttribute
