+++ src/Grpc.Extension.Abstract/IGrpcService.cs
@@ -5,10 +5,15 @@ using System.Text;
 namespace Grpc.Extension.Abstract
 {
     /// <summary>
-    /// GrpcService
+    /// GrpcService(CodeFirst)
     /// </summary>
     public interface IGrpcService
     {
 
     }
+
+    /// <summary>
+    /// 非Grpc方法
+    /// </summary>
+    public sealed class NotGrpcMethodAttribute : Attribute { }
 }
